+++ src/pages/Analysis.vue
@@ -1,17 +1,7 @@
 <template>
   <div class="container py-4">
     <!-- 기간 선택 -->
-    <div class="mb-4 d-flex flex-wrap gap-2">
-      <button
-        v-for="m in [1, 3, 6, 12]"
-        :key="m"
-        class="btn"
-        :class="store.period === m ? 'btn-primary' : 'btn-secondary'"
-        @click="selectPeriod(m)"
-      >
-        {{ m }}개월
-      </button>
-    </div>
+    <selectPeriod />
 
     <!-- 요약 및 점수 -->
     <div class="row g-4">
@@ -23,58 +13,19 @@
 
             <!-- 총 수입, 지출 등 -->
             <div class="row">
-              <div class="col-12 col-md-3 mb-3">
-                <small class="text-muted">
-                  총 수입
-                  <span
-                    :class="
-                      parseFloat(store.periodStat.diffIncome.replace('%', '')) >
-                      0
-                        ? 'text-success'
-                        : 'text-danger'
-                    "
-                  >
-                    {{ store.periodStat.diffIncome }}
-                  </span>
-                </small>
-                <div class="h5">
-                  <AnimatedNumber :value="store.periodStat.income" />원
-                </div>
-                <div class="small">
-                  {{
-                    store.periodStat.prevIncome === 0
-                      ? '지난 데이터 없음'
-                      : `지난 수입 ${store.periodStat.prevIncome.toLocaleString()}원`
-                  }}
-                </div>
-              </div>
+              <analysisCard
+                :diffData="store.periodStat.diffIncome"
+                :data="store.periodStat.income"
+                :prevData="store.periodStat.prevIncome"
+                :title="'수입'"
+              />
 
-              <div class="col-12 col-md-3 mb-3">
-                <small class="text-muted">
-                  총 지출
-                  <span
-                    :class="
-                      parseFloat(
-                        store.periodStat.diffExpense.replace('%', '')
-                      ) < 0
-                        ? 'text-success'
-                        : 'text-danger'
-                    "
-                  >
-                    {{ store.periodStat.diffExpense }}
-                  </span>
-                </small>
-                <div class="h5">
-                  <AnimatedNumber :value="store.periodStat.expense" />원
-                </div>
-                <div class="small">
-                  {{
-                    store.periodStat.prevExpense === 0
-                      ? '지난 데이터 없음'
-                      : `지난 지출 ${store.periodStat.prevExpense.toLocaleString()}원`
-                  }}
-                </div>
-              </div>
+              <analysisCard
+                :diffData="store.periodStat.diffExpense"
+                :data="store.periodStat.expense"
+                :prevData="store.periodStat.prevExpense"
+                :title="'지출'"
+              />
 
               <div class="col-12 col-md-3 mb-3">
                 <small class="text-muted">합계</small>
@@ -104,28 +55,7 @@
       </div>
 
       <!-- 소비 점수 -->
-      <div
-        class="col-12 col-md-4 d-flex justify-content-center align-items-center"
-      >
-        <div class="score-circle text-center">
-          <div
-            class="circle"
-            :style="{
-              borderColor: borderColor,
-              borderWidth: '10px',
-              borderStyle: 'solid',
-            }"
-          >
-            <span class="score" :class="gradeInfo.color">
-              {{ gradeInfo.grade }}
-            </span>
-          </div>
-          <div class="mt-2" :class="gradeInfo.color">
-            <div class="fw-bold">{{ gradeInfo.grade }} 등급</div>
-            <div class="text-muted">{{ gradeInfo.message }}</div>
-          </div>
-        </div>
-      </div>
+      <scoreCircle :expense="periodStat.expense" :income="periodStat.income" />
     </div>
 
     <!-- 차트 -->
@@ -138,7 +68,6 @@
           </div>
         </div>
       </div>
-
       <div class="col-12 col-md-8">
         <div class="card h-100">
           <div class="card-body">
@@ -147,7 +76,6 @@
               <span class="text-muted" id="chart-period-label"
                 >데이터 없음</span
               >
-              <!-- 또는 <select> 요소 사용도 가능 -->
             </div>
             <canvas ref="weeklyChart"></canvas>
           </div>
@@ -160,58 +88,18 @@
 <script setup>
 import { ref, computed, onMounted, watch } from 'vue';
 import { useAnalysisStore } from '@/stores/analysisStore';
-import { renderCategoryChart, renderWeeklyChart } from '@/composables/useChart';
+import { renderCategoryChart, renderWeeklyChart } from '@/services/useChart.js';
 import AnimatedNumber from '@/components/AnimatedNumber.vue';
+import selectPeriod from '@/components/analysis/selectPeriod.vue';
+import analysisCard from '@/components/analysis/analysisCard.vue';
+import scoreCircle from '@/components/analysis/scoreCircle.vue';
 
 const categoryChart = ref(null);
 const weeklyChart = ref(null);
 
 // ✅ Pinia 스토어 가져오기
 const store = useAnalysisStore();
-const { pepriod, periodStat, fetchUserTransactions, selectPeriod, filters } =
-  store;
-
-const borderColor = computed(() => {
-  const colorMap = {
-    'text-success': '#2e7d32',
-    'text-primary': '#1976d2',
-    'text-warning': '#fbc02d',
-    'text-danger': '#d32f2f',
-  };
-  return colorMap[gradeInfo.value.color] || '#ccc';
-});
-
-const gradeInfo = computed(() => {
-  const val = (periodStat.expense / periodStat.income) * 100;
-  let grade = '',
-    flag = '',
-    color = '';
-  if (val <= 30) {
-    grade = 'A';
-    flag = '30%';
-    color = 'text-success';
-  } else if (val <= 50) {
-    grade = 'B';
-    flag = '50%';
-    color = 'text-primary';
-  } else if (val <= 80) {
-    grade = 'C';
-    flag = '80%';
-    color = 'text-warning';
-  } else {
-    grade = 'D';
-    flag = '80%';
-    color = 'text-danger';
-  }
-  return {
-    grade,
-    message:
-      grade === 'D'
-        ? `수입의 ${flag}이상 지출했어요!`
-        : `수입의 ${flag}정도 지출했어요!`,
-    color,
-  };
-});
+const { periodStat, fetchUserTransactions, filters } = store;
 
 onMounted(async () => {
   await fetchUserTransactions();
@@ -243,38 +131,6 @@ watch(
 </script>
 
 <style scoped>
-.score-circle .circle {
-  width: 100px;
-  height: 100px;
-  border: 10px solid #385531;
-  border-radius: 50%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin: auto;
-}
-.score-circle .score {
-  font-size: 1.5rem;
-  color: #6f42c1;
-  font-weight: bold;
-}
-
-.border-success {
-  border-color: #2e7d32;
-}
-.border-primary {
-  border-color: #1976d2;
-}
-.border-warning {
-  border-color: #fbc02d;
-}
-.border-danger {
-  border-color: #d32f2f;
-}
-.btn-secondary {
-  --bs-btn-bg: #a5abaf;
-  --bs-btn-border-color: #a5abaf;
-}
 canvas {
   width: 100% !important;
   height: 400px !important;
